Added some safe checking and moved the theme for the DatePicker component into a separate file for less clutter
