+++ src/components/CustomDatePicker/CustomDatePicker.js
@@ -3,33 +3,9 @@ import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import TextField from "@mui/material/TextField";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { ThemeProvider } from "@mui/material/styles";
 
-const theme = createTheme({
-  components: {
-    MuiIconButton: {
-      styleOverrides: {
-        sizeMedium: {
-          color: '#2C3E50'
-        }
-      }
-    },
-    MuiOutlinedInput: {
-      styleOverrides: {
-        root: {
-          color: '#2C3E50'
-        }
-      }
-    },
-    MuiInputLabel: {
-      styleOverrides: {
-        root: {
-          color: '#2C3E50'
-        }
-      }
-    }
-  }
-});
+import theme from './styles';
 
 const CustomDatePicker = ({ date, setDate }) => {
