+++ src/components/RoverPhoto/RoverPhoto.js
@@ -4,7 +4,7 @@ import { Card, CardMedia } from "@mui/material";
 const RoverPhoto = ({ url }) => {
     return (
         <Card>
-            <CardMedia component="img" src={url} sx={{ height: '300px'}}/>
+            <CardMedia component="img" src={url} height='300px'/>
         </Card>
     )
 };
